pyproject.toml
[project]
name = "ai-tutorial"
version = "0.1.0"
description = "Interactive AI Engineering Tutorial Website - Learn LLMs, RAG, Agents, MCP and more"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    # LLM Providers
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    # Vector Database
    "chromadb>=0.4.22",
    # Utilities
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "markdown>=3.5.2",
    "pygments>=2.17.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_tutorial"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
